Added InputFIeld.bf and modified main menu.
Main menu now has an input field that appears when clicking the start button once. It'll start the game when clicking it again.

Purpose of the input field is to allow the user to input the amount of boids they want, though that functionality is not currently implemented.

TODO:
Remove the hardcoding in InputFIeld.bf that makes it only accept numbers

Create a UI element that inherits from InputFIeld.bf that is specifically for only having number inputs.